# apps/api/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/invoices/{id}": {
      "get": {
        "operationId": "InvoicesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "6630f0c8a1b2c3d4e5f6a7b8",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid X-Api-Key header"
          },
          "404": {
            "description": "Invoice not found"
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Get invoice by id",
        "tags": [
          "invoices"
        ]
      }
    },
...
    "/v1/admin/invoices/{id}": {
      "get": {
        "operationId": "AdminController_getInvoice",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid X-Admin-Api-Key header"
          }
        },
        "security": [
          {
            "admin-key": []
          }
        ],
        "summary": "List invoices",
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Payments API",
    "description": "REST API for payment processor",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      },
      "admin-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Api-Key"
      },
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "InvoiceResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "6630f0c8a1b2c3d4e5f6a7b8"
          },
          "chain": {
            "type": "string",
            "enum": [
              "xmr",
              "firo"
            ],
            "example": "xmr"
          },
          "asset": {
            "type": "string",
            "enum": [
              "xmr",
              "firo"
            ],
            "example": "xmr"
          },
          "assetDecimals": {
            "type": "number",
            "example": 12,
            "description": "Decimals for atomic conversion"
          },
          "address": {
            "type": "string",
            "example": "5A1jR...subaddress...xyz",
            "description": "Unique receiving address for this invoice"
          },
          "addressIndex": {
            "type": "number",
            "example": 1,
            "description": "Address index (XMR subaddress index or HD derivation index)"
          },
          "amountAtomic": {
            "type": "string",
            "example": "123456789012",
            "description": "Amount owed in atomic units (string)"
          },
          "amountFiat": {
            "type": "number",
            "example": 19.99
          },
          "fiatCurrency": {
            "type": "string",
            "example": "USD"
          },
          "rate": {
            "type": "number",
            "example": 0.00617,
            "description": "Asset per 1 unit fiat at lock time"
          },
          "rateLockedAt": {
            "type": "string",
            "example": "2026-05-06T12:34:56.000Z"
          },
          "expiresAt": {
            "type": "string",
            "example": "2026-05-06T12:54:56.000Z"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "seen",
              "confirmed",
              "underpaid",
              "expired",
              "cancelled"
            ],
            "example": "pending"
          },
          "confirmationsRequired": {
            "type": "number",
            "example": 1
          },
          "confirmations": {
            "type": "number",
            "example": 0
          },
          "receivedAtomic": {
            "type": "string",
            "example": "0",
            "description": "Total atomic units received across all payments"
          },
          "firstSeenAt": {
            "type": "string",
            "example": "2026-05-06T12:36:01.000Z"
          },
          "paidAt": {
            "type": "string",
            "example": "2026-05-06T12:38:14.000Z"
          },
          "webhookUrl": {
            "type": "string",
            "example": "https://merchant.example.com/hook"
          },
          "chainData": {
            "type": "object",
            "additionalProperties": true,
            "description": "Chain-specific data (e.g. txHash, blockNumber for EVM)"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          },
          "createdAt": {
            "type": "string",
            "example": "2026-05-06T12:34:56.000Z"
          }
        },
        "required": [
          "id",
          "chain",
          "asset",
          "assetDecimals",
          "address",
          "addressIndex",
          "amountAtomic",
          "amountFiat",
          "fiatCurrency",
          "rate",
          "rateLockedAt",
          "expiresAt",
          "status",
          "confirmationsRequired",
          "confirmations",
          "receivedAtomic",
          "createdAt"
        ]
      },
...
    }
  }
}